Add TestMemStore
