+++ src/kv/test-mem-store.ts
@@ -1,0 +1,11 @@
+import {MemStore} from './mem-store';
+import type {Value} from './store';
+
+export class TestMemStore extends MemStore {
+  /**
+   * This exposes the underlying map for testing purposes.
+   */
+  entries(): Iterable<[string, Value]> {
+    return this._map.entries();
+  }
+}
