+++ src/migrate/migrate.test.ts
@@ -1,5 +1,4 @@
 import {expect} from '@esm-bundle/chai';
-import {MemStore} from '../kv/mod';
 import {initHasher} from '../hash';
 import {
   chatSampleV0,
@@ -10,7 +9,8 @@ import {
   testIndexDataV1,
 } from './migrate-sample-data';
 import {migrate} from './migrate';
-import type {Value} from '../kv/store';
+import type {Store, Value} from '../kv/store';
+import {TestMemStore} from '../kv/test-mem-store';
 
 setup(async () => {
   await initHasher();
@@ -20,13 +20,12 @@ async function testMigrate(
   inputdata: Record<string, Value>,
   expected: Record<string, Value>,
 ): Promise<void> {
-  const kv = new MemStore();
+  const kv = new TestMemStore();
   await writeSampleData(kv, inputdata);
 
   await migrate(kv);
 
-  // @ts-expect-error Using private property.
-  const actual = Object.fromEntries(kv._map.entries());
+  const actual = Object.fromEntries(kv.entries());
   expect(actual).to.deep.equal(expected);
 }
 
@@ -38,12 +37,12 @@ test('test data sample', async () => {
   await testMigrate(testDataV0, testDataV1);
 });
 
-test.only('test data sample with index', async () => {
+test('test data sample with index', async () => {
   await testMigrate(testIndexDataV0, testIndexDataV1);
 });
 
 async function writeSampleData(
-  kv: MemStore,
+  kv: Store,
   data: Record<string, Value>,
 ): Promise<void> {
   return kv.withWrite(async w => {
